backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv==1.2.1
pymongo==4.15.5